CLJ-1107: Throw exception for 'get' called on unsupported types
Before this patch, clojure.core/get defaults to returning nil when it
is called on a type which it does not support.

After this patch, clojure.core/get throws an exception if it is called
on a non-Associative type. This provides more visibility for common
mistakes such as (:foo r) where r is a Ref or Atom containing a map.

This change uncovered some incorrect syntax in a Clojure test file for
multimethods.

One extra change was needed in core_deftype.clj to return an
appropriate error message in clojure.core/extend.