Protocol, interface method declarations don't allow for varags.
Currently, for example

  (defprotocol FooBar
    (foo [this & more]))

compiles just fine, and & is interpreted as a usual argument that happens to be
named & without special meaning.  But clearly, the user wanted to specify a
varags parameter here.  The same applies to definterface.

Similarly, providing method implementations via defrecord, deftype, and reify
don't allow for varags (but dynamic extenions via extend do).

So this patch makes defprotocol and definterface throw an Exception if a user
tries to use varargs in method signatures.  Similarly, defrecord, deftype, and
reify throw an Exception if any method implementation arglist contains a
varargs argument.